requirements.txt
essentia-tensorflow
matplotlib
numpy

